feat: more than 7 modules
