+++ archiv.js
@@ -1,0 +1,40 @@
+import fs from 'node:fs/promises'
+import path from 'path'
+import { createReadStream, createWriteStream } from "node:fs"
+import zlib from 'node:zlib'
+
+export async function compress (path_to_file, path_to_destination) {
+  await compresser(path_to_file, path_to_destination, zlib.createBrotliCompress())
+}
+
+export async function decompress (path_to_file, path_to_destination) {
+  await compresser(path_to_file, path_to_destination, zlib.createBrotliDecompress())
+}
+
+async function compresser(path_to_file, new_filename, brotli) {
+  const oldpath = path.resolve(path_to_file)
+  const newpath = path.resolve(new_filename)
+  try {
+    await fs.stat(oldpath)
+    try {
+      await fs.stat(newpath)
+      console.log('Operation failed')
+    } catch (err) {
+      if (err.code === 'ENOENT') {
+        await new Promise((resolve, reject) => {
+          const inStream = createReadStream(oldpath)
+          const outStream = createWriteStream(newpath)
+          inStream.pipe(brotli).pipe(outStream)
+          inStream.on('error', reject)
+          outStream.on('error', reject)
+          outStream.on('finish', resolve)
+        })
+        await rm(path_to_file)
+      } else {
+        console.log('Operation failed')
+      }
+    }
+  } catch {
+    console.log('Operation failed')
+  }
+}
