+++ fs/index.js
@@ -1,0 +1,7 @@
+export * from './add.js'
+export * from './cat.js'
+export * from './cp.js'
+export * from './mkdir.js'
+export * from './mv.js'
+export * from './rm.js'
+export * from './rn.js'
